Basic application framework setup
UI with list and start button setup.

folder scanner searches all sub folders breadth first and
prints them to debug out.

Cludge included in the form of a timer because the search slot
emitting a signal linked to it's self didn't give the UI time
to do other stuff and ended with a SEGFAULT.

The better approach seems to be a worker thread.