Cludge for the timer cludge
I head from an unreliable source that setting the interval to 0 on a timer
tells the event loop to run the function on the next loop.

I think the worker thread is still the right option, but this is good enough
for now; I'll leave it here and get the rest of the functionallity going
before I come back and look at it again.